requirements.txt
asgiref==3.6.0
Django==4.1.6
django-braces==1.15.0
django-crispy-forms==1.14.0
sqlparse==0.4.3
